lint

+++ pubky/src/common.rs
@@ -18,5 +18,5 @@ pub(crate) fn timestamp() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_secs() as u64
+        .as_secs()
 }
